Refactor agent-browser skill: fix bugs, deduplicate scripts, slim docs
- Extract shared utilities into scripts/_lib.py; both scripts import from
  it, eliminating ~200 lines of duplication and fixing the yaml_scalar bug
  where [, ], {, } were not quoted in check-site-experience.py output
- Add --site flag to check-site-experience.py for direct host lookup
  without a running browser; make --site required=True in validator
- Remove skill_root absolute path from validate output; remove dead assert
- Trim SKILL.md maintenance section from 36 to ~15 lines: keep trigger,
  pointer to experience-authoring.md, skip criteria, and validation command;
  move detailed artifact/escalation guidance exclusively to the reference doc
- Add "run from the installed skill root" note to script command examples
- Remove hardcoded dev-repo path from SKILL.md runtime maintenance section
- Update tests: drop phrase-locking assertions that enforced overlap between
  SKILL.md and experience-authoring.md; keep constraint-value checks only
