+++ tests/test_site_experience_validation.py
@@ -188,31 +188,15 @@ class SiteExperienceMaintenanceDocumentationTest(unittest.TestCase):
         self.assertIn("maintenance check", skill_text)
         self.assertIn("before deciding", skill_text)
         self.assertIn("maintaining site experience", skill_text)
-        self.assertIn("update or generalize the closest fit", skill_text)
-        self.assertIn("Tool and workflow budgets are limited", skill_text)
         self.assertIn("Final answer: list maintained paths or the skip reason", skill_text)
         self.assertIn("Maintain mechanics, not content", skill_text)
-        self.assertLessEqual(len(skill_text.splitlines()), 110)
-        self.assertIn("reusable", skill_text)
-        self.assertIn("workflow", skill_text)
         self.assertIn("Ask before writing only", skill_text)
         self.assertIn("validate-site-experience.py", skill_text)
+        self.assertLessEqual(len(skill_text.splitlines()), 110)
 
     def test_authoring_guide_records_maintenance_constraints(self):
         guide_text = AUTHORING_GUIDE.read_text(encoding="utf-8")
 
-        self.assertIn("apply these gates", guide_text)
-        self.assertIn("write/update without asking", guide_text)
-        self.assertIn("report the skip reason", guide_text)
-        self.assertIn("Keep tools atomic", guide_text)
-        self.assertIn("Private or authenticated sessions are allowed", guide_text)
-        self.assertIn("Use existing artifacts first", guide_text)
-        self.assertIn("update or generalize the closest fit", guide_text)
-        self.assertIn("add a new artifact only when the mechanic has no clear home", guide_text)
-        self.assertIn("Tool and workflow budgets are limited", guide_text)
-        self.assertIn("Tool: reusable automation", guide_text)
-        self.assertIn("Workflow: non-obvious sequence", guide_text)
-        self.assertIn("Site note: stable site knowledge", guide_text)
         self.assertIn("2000 characters", guide_text)
         self.assertIn("20 files", guide_text)
         self.assertIn("Do not save `@eN` refs", guide_text)
